# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gameng API
  version: 0.0.1
  description: Server-side data-driven RPG engine API
servers:
  - url: http://localhost:3000
...
paths:
...
  /{gameInstanceId}/state/player/{playerId}:
    get:
      operationId: getPlayerState
      summary: Get player state
      description: Returns the full player sub-object from the game state, including characters and gear. Requires authentication and ownership of the player.
      tags:
        - state
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gameInstanceId"
        - $ref: "#/components/parameters/playerId"
      responses:
        "200":
          description: Player state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
              example:
                characters:
                  char_1:
                    classId: warrior
                    level: 3
                    equipped:
                      right_hand: gear_1
                gear:
                  gear_1:
                    gearDefId: sword_basic
                    level: 2
                    equippedBy: char_1
        "404":
          description: Game instance or player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                instance_not_found:
                  summary: Game instance not found
                  value:
                    errorCode: INSTANCE_NOT_FOUND
                    errorMessage: "Game instance 'unknown' not found."
                player_not_found:
                  summary: Player not found
                  value:
                    errorCode: PLAYER_NOT_FOUND
                    errorMessage: "Player 'nobody' not found."
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                errorCode: UNAUTHORIZED
                errorMessage: "Missing or invalid Bearer token."
        "403":
          description: Authenticated actor does not own the target player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                errorCode: OWNERSHIP_VIOLATION
                errorMessage: "Actor does not own player 'player_1'."
  /{gameInstanceId}/character/{characterId}/stats:
    get:
      operationId: getCharacterStats
      summary: Get computed character stats
      description: Returns the final calculated stats for a character, combining base class stats, gear, set bonuses, and growth scaling. Requires authentication and ownership of the character's player.
      tags:
        - state
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/gameInstanceId"
        - $ref: "#/components/parameters/characterId"
      responses:
        "200":
          description: Computed character stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterStats"
              example:
                characterId: char_1
                classId: warrior
                level: 3
                finalStats:
                  strength: 12.5
                  hp: 45
        "404":
          description: Game instance or character not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                instance_not_found:
                  summary: Game instance not found
                  value:
                    errorCode: INSTANCE_NOT_FOUND
                    errorMessage: "Game instance 'unknown' not found."
                character_not_found:
                  summary: Character not found
                  value:
                    errorCode: CHARACTER_NOT_FOUND
                    errorMessage: "Character 'char_99' not found."
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                errorCode: UNAUTHORIZED
                errorMessage: "Missing or invalid Bearer token."
        "403":
          description: Authenticated actor does not own the character's player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                errorCode: OWNERSHIP_VIOLATION
                errorMessage: "Actor does not own player 'player_1'."
components:
  parameters:
    gameInstanceId:
      name: gameInstanceId
      in: path
      required: true
      description: Unique identifier for the game instance
      schema:
        type: string
        minLength: 1
    playerId:
      name: playerId
      in: path
      required: true
      description: Unique identifier for the player
      schema:
        type: string
        minLength: 1
    characterId:
      name: characterId
      in: path
      required: true
      description: Unique identifier for the character
      schema:
        type: string
        minLength: 1
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          minLength: 1
        errorMessage:
          type: string

    # -- Config domain --
    StatMap:
      type: object
      additionalProperties:
        type: number
      description: Map of statId to numeric value. Missing stats default to 0.
...
    Character:
      type: object
      required:
        - classId
        - level
        - equipped
      additionalProperties: false
      properties:
        classId:
          type: string
          minLength: 1
        level:
          type: integer
          minimum: 1
        equipped:
          type: object
          additionalProperties:
            type: string
            minLength: 1
          description: Map of slotId to gearId. Empty slots are absent.
    GearInstance:
      type: object
      required:
        - gearDefId
        - level
      additionalProperties: false
      properties:
        gearDefId:
          type: string
          minLength: 1
        level:
          type: integer
          minimum: 1
        equippedBy:
          type:
            - string
            - "null"
          description: characterId if equipped, null or absent if in inventory.
    Player:
      type: object
      required:
        - characters
        - gear
      additionalProperties: false
      properties:
        characters:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Character"
          description: Map of characterId to character data.
        gear:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/GearInstance"
          description: Map of gearId to gear instance data (inventory + equipped).
...
    CharacterStats:
      type: object
      required:
        - characterId
        - classId
        - level
        - finalStats
      additionalProperties: false
      properties:
        characterId:
          type: string
          minLength: 1
        classId:
          type: string
          minLength: 1
        level:
          type: integer
          minimum: 1
        finalStats:
          $ref: "#/components/schemas/StatMap"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer token for authentication. For CreateActor transactions, use the
        ADMIN_API_KEY. For all other endpoints, use the actor's apiKey obtained
        via CreateActor.